Added defensive code to handle apps that call Ctl functions on non-Control objects, like CtlSetUsable() and CtlSetLabel().

+++ src/PalmOS/Core/UI/Form.h
@@ -211,7 +211,7 @@ typedef struct FormLabelTag
 	UInt8 						reserved;
 	Char *						text;
 
-  UInt16 marker;
+  UInt32 marker;
 	PointType extent;
   Int16 len;
   Char buf[0];
